Refactor: test 도메인 별로 파일 나눔
